Split out base Character Edit form
