+++ frontend/src/components/molecules/EditCharacterForm.tsx
@@ -1,12 +1,8 @@
-import Button from '@material-ui/core/Button';
-import TextField from '@material-ui/core/TextField';
-import { ChangeEvent } from 'react';
 import * as React from 'react';
 import { connect } from 'react-redux';
 
 import { characterActions } from '../../store/actions';
-import { Paper, Typography } from '@material-ui/core';
-import { DeleteButton } from '../atoms/DeleteButton';
+import { BaseEditCharacterForm } from './BaseEditCharacterForm';
 
 interface Props {
     characterId: string;
@@ -23,30 +19,16 @@ interface MapProps {
 
 type AllProps = Props & MapProps;
 
-interface State {
-    name: string;
-    description: string;
-}
-
-class EditCharacterFormRaw extends React.Component<AllProps, State> {
+class EditCharacterFormRaw extends React.Component<AllProps> {
     constructor(props: AllProps) {
         super(props);
-        this.state = { name: props.initialName, description: props.initialDescription };
     }
 
-    handleChangeName = (event: ChangeEvent<HTMLInputElement>) => {
-        this.setState({ name: event.target.value });
-    };
-
-    handleChangeDescription = (event: ChangeEvent<HTMLInputElement>) => {
-        this.setState({ description: event.target.value });
-    };
-
-    handleSubmit = () => {
+    handleSubmit = (name: string, description: string) => {
         this.props.editCharacter({
             characterId: this.props.characterId,
-            name: this.state.name,
-            description: this.state.description
+            name,
+            description
         });
         if (this.props.onSubmitComplete) {
             this.props.onSubmitComplete();
@@ -60,37 +42,14 @@ class EditCharacterFormRaw extends React.Component<AllProps, State> {
         }
     };
 
-    canSubmit = () => this.state.name && this.state.description;
-
-    render() {
-        return (
-            <Paper className="modalPaper">
-                <div className="modalContainer">
-                    <div className="modalHeader">
-                        <Typography variant="h5">Update character</Typography>
-                        <DeleteButton onConfirm={this.handleDelete} />
-                    </div>
-                    <TextField label="Name" value={this.state.name} onChange={this.handleChangeName} margin="dense" />
-                    <TextField
-                        multiline
-                        variant="outlined"
-                        label="Description"
-                        value={this.state.description}
-                        onChange={this.handleChangeDescription}
-                        margin="normal"
-                    />
-                    <Button
-                        variant="contained"
-                        color="primary"
-                        onClick={this.handleSubmit}
-                        disabled={!this.canSubmit()}
-                    >
-                        Update
-                    </Button>
-                </div>
-            </Paper>
-        );
-    }
+    render = () => (
+        <BaseEditCharacterForm
+            initialName={this.props.initialName}
+            initialDescription={this.props.initialDescription}
+            onSubmit={this.handleSubmit}
+            onDelete={this.handleDelete}
+        />
+    );
 }
 
 export const EditCharacterForm = connect(null, {
